remove more unused filtering
