+++ js_modules/dagster-ui/packages/ui-core/src/asset-selection/useAssetSelectionState.oss.tsx
@@ -1,5 +1,9 @@
 import {useQueryPersistedState} from '../hooks/useQueryPersistedState';
 
+export function getAssetSelectionQueryString(search: string = location.search) {
+  const params = new URLSearchParams(search);
+  return `&asset-selection=${params.get('asset-selection')}`;
+}
 export function useAssetSelectionState() {
   return useQueryPersistedState<string>({
     queryKey: 'asset-selection',
